python/requirements.txt
sympy>=1.13.3
fastapi==0.115.0
uvicorn==0.30.6
gradio==4.44.0
numpy==2.1.1
pydantic>=2.11.7
pint==0.24.4
antlr4-python3-runtime==4.11.1
httpx==0.28.1
torch==2.8.0
transformers>=4.55.2
peft>=0.3.0
pyyaml==6.0.2
datasets==2.21.0
trl>=0.7.4
vllm==0.11.0